Move Event to the common crate
