+++ mio-common/src/event.rs
@@ -1,0 +1,90 @@
+use crate::{Ready, Token};
+
+/// An readiness event returned by [`Poll::poll`].
+///
+/// `Event` is a [readiness state] paired with a [`Token`]. It is returned by
+/// [`Poll::poll`].
+///
+/// For more documentation on polling and events, see [`Poll`].
+///
+/// # Examples
+///
+/// ```
+/// use mio::{Ready, Token};
+/// use mio::event::Event;
+///
+/// let event = Event::new(Ready::readable() | Ready::writable(), Token(0));
+///
+/// assert_eq!(event.readiness(), Ready::readable() | Ready::writable());
+/// assert_eq!(event.token(), Token(0));
+/// ```
+///
+/// [`Poll::poll`]: ../struct.Poll.html#method.poll
+/// [`Poll`]: ../struct.Poll.html
+/// [readiness state]: ../struct.Ready.html
+/// [`Token`]: ../struct.Token.html
+#[derive(Copy, Clone, Eq, PartialEq, Debug)]
+pub struct Event {
+    kind: Ready,
+    token: Token,
+}
+
+impl Event {
+    /// Creates a new `Event` containing `readiness` and `token`
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use mio::{Ready, Token};
+    /// use mio::event::Event;
+    ///
+    /// let event = Event::new(Ready::readable() | Ready::writable(), Token(0));
+    ///
+    /// assert_eq!(event.readiness(), Ready::readable() | Ready::writable());
+    /// assert_eq!(event.token(), Token(0));
+    /// ```
+    pub fn new(readiness: Ready, token: Token) -> Event {
+        Event {
+            kind: readiness,
+            token,
+        }
+    }
+
+    /// Returns the event's readiness.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use mio::{Ready, Token};
+    /// use mio::event::Event;
+    ///
+    /// let event = Event::new(Ready::readable() | Ready::writable(), Token(0));
+    ///
+    /// assert_eq!(event.readiness(), Ready::readable() | Ready::writable());
+    /// ```
+    pub fn readiness(&self) -> Ready {
+        self.kind
+    }
+
+    // FIXME(Thomas): remove.
+    #[doc(hidden)]
+    pub fn readiness_mut(&mut self) -> &mut Ready {
+        &mut self.kind
+    }
+
+    /// Returns the event's token.
+    ///
+    /// # Examples
+    ///
+    /// ```
+    /// use mio::{Ready, Token};
+    /// use mio::event::Event;
+    ///
+    /// let event = Event::new(Ready::readable() | Ready::writable(), Token(0));
+    ///
+    /// assert_eq!(event.token(), Token(0));
+    /// ```
+    pub fn token(&self) -> Token {
+        self.token
+    }
+}
